rename "opener" to "snapfile"
